+++ app/ide-desktop/lib/dashboard/src/layouts/AssetPanel.tsx
@@ -7,6 +7,7 @@ import * as textProvider from '#/providers/TextProvider'
 import type * as assetEvent from '#/events/assetEvent'
 import type * as assetListEvent from '#/events/assetListEvent'
 
+import AssetProjectSessions from '#/layouts/AssetProjectSessions'
 import AssetProperties from '#/layouts/AssetProperties'
 import AssetVersions from '#/layouts/AssetVersions/AssetVersions'
 import type Category from '#/layouts/CategorySwitcher/Category'
@@ -29,6 +30,7 @@ import * as tailwindMerge from '#/utilities/tailwindMerge'
 enum AssetPanelTab {
   properties = 'properties',
   versions = 'versions',
+  projectSessions = 'projectSessions',
 }
 
 // ============================
@@ -84,6 +86,11 @@ export default function AssetPanel(props: AssetPanelProps) {
       savedTab === AssetPanelTab.versions
     ) {
       return AssetPanelTab.properties
+    } else if (
+      item?.item.type !== backendModule.AssetType.project &&
+      savedTab === AssetPanelTab.projectSessions
+    ) {
+      return AssetPanelTab.properties
     } else {
       return savedTab
     }
@@ -109,16 +116,16 @@ export default function AssetPanel(props: AssetPanelProps) {
         event.stopPropagation()
       }}
     >
-      <ariaComponents.ButtonGroup className="mt-4">
+      <ariaComponents.ButtonGroup className="mt-4 grow-0 basis-8">
         {item != null &&
           item.item.type !== backendModule.AssetType.secret &&
           item.item.type !== backendModule.AssetType.directory && (
             <ariaComponents.Button
-              size="custom"
-              variant="custom"
+              size="medium"
+              variant="ghost"
               className={tailwindMerge.twMerge(
-                'button pointer-events-auto h-8 select-none bg-frame px-button-x leading-cozy transition-colors hover:bg-primary/[8%]',
-                tab === AssetPanelTab.versions && 'bg-primary/[8%] active'
+                'pointer-events-auto disabled:opacity-100',
+                tab === AssetPanelTab.versions && 'bg-white opacity-100'
               )}
               onPress={() => {
                 setTab(oldTab =>
@@ -131,6 +138,28 @@ export default function AssetPanel(props: AssetPanelProps) {
               {getText('versions')}
             </ariaComponents.Button>
           )}
+        {item != null && item.item.type === backendModule.AssetType.project && (
+          <ariaComponents.Button
+            size="medium"
+            variant="ghost"
+            isDisabled={tab === AssetPanelTab.projectSessions}
+            className={tailwindMerge.twMerge(
+              'pointer-events-auto disabled:opacity-100',
+              tab === AssetPanelTab.projectSessions && 'bg-white opacity-100'
+            )}
+            onPress={() => {
+              setTab(oldTab =>
+                oldTab === AssetPanelTab.projectSessions
+                  ? AssetPanelTab.properties
+                  : AssetPanelTab.projectSessions
+              )
+            }}
+          >
+            {getText('projectSessions')}
+          </ariaComponents.Button>
+        )}
+        {/* Spacing. The top right asset and user bars overlap this area. */}
+        <div className="grow" />
       </ariaComponents.ButtonGroup>
       {item == null || setItem == null || backend == null ? (
         <div className="grid grow place-items-center text-lg">
@@ -155,6 +184,10 @@ export default function AssetPanel(props: AssetPanelProps) {
               dispatchAssetListEvent={dispatchAssetListEvent}
             />
           )}
+          {tab === AssetPanelTab.projectSessions &&
+            item.type === backendModule.AssetType.project && (
+              <AssetProjectSessions backend={backend} item={item} />
+            )}
         </>
       )}
     </div>