+++ app/ide-desktop/lib/dashboard/src/layouts/AssetsTable.tsx
@@ -7,7 +7,6 @@ import DropFilesImage from 'enso-assets/drop_files.svg'
 
 import * as mimeTypes from '#/data/mimeTypes'
 
-import * as asyncEffectHooks from '#/hooks/asyncEffectHooks'
 import * as backendHooks from '#/hooks/backendHooks'
 import * as eventHooks from '#/hooks/eventHooks'
 import * as intersectionHooks from '#/hooks/intersectionHooks'
@@ -214,7 +213,7 @@ function insertAssetTreeNodeChildren(
     node => node.item.type !== backendModule.AssetType.specialEmpty
   )
   const nodesToInsert = children.map(asset =>
-    AssetTreeNode.fromAsset(asset, directoryKey, directoryId, depth)
+    AssetTreeNode.fromAsset(asset, directoryKey, directoryId, depth, `${item.path}/${asset.title}`)
   )
   const newNodes = array.splicedBefore(
     nodes,
@@ -260,6 +259,7 @@ function insertArbitraryAssetTreeNodeChildren(
           directoryKey,
           directoryId,
           depth,
+          `${item.path}/${asset.title}`,
           getKey?.(asset) ?? asset.id
         )
       )
@@ -329,12 +329,8 @@ export interface AssetsTableState {
     title?: string | null,
     override?: boolean
   ) => void
-  readonly doOpenEditor: (
-    project: backendModule.ProjectAsset,
-    setProject: React.Dispatch<React.SetStateAction<backendModule.ProjectAsset>>,
-    switchPage: boolean
-  ) => void
-  readonly doCloseEditor: (project: backendModule.ProjectAsset) => void
+  readonly doOpenEditor: () => void
+  readonly doCloseEditor: (projectId: backendModule.ProjectId) => void
   readonly doCopy: () => void
   readonly doCut: () => void
   readonly doPaste: (
@@ -361,7 +357,6 @@ export interface AssetsTableProps {
   readonly category: Category
   readonly setSuggestions: (suggestions: assetSearchBar.Suggestion[]) => void
   readonly initialProjectName: string | null
-  readonly projectStartupInfo: backendModule.ProjectStartupInfo | null
   readonly assetListEvents: assetListEvent.AssetListEvent[]
   readonly dispatchAssetListEvent: (event: assetListEvent.AssetListEvent) => void
   readonly assetEvents: assetEvent.AssetEvent[]
@@ -369,21 +364,16 @@ export interface AssetsTableProps {
   readonly setAssetPanelProps: (props: assetPanel.AssetPanelRequiredProps | null) => void
   readonly setIsAssetPanelTemporarilyVisible: (visible: boolean) => void
   readonly targetDirectoryNodeRef: React.MutableRefObject<assetTreeNode.AnyAssetTreeNode<backendModule.DirectoryAsset> | null>
-  readonly doOpenEditor: (
-    backend: Backend,
-    project: backendModule.ProjectAsset,
-    setProject: React.Dispatch<React.SetStateAction<backendModule.ProjectAsset>>,
-    switchPage: boolean
-  ) => void
-  readonly doCloseEditor: (project: backendModule.ProjectAsset) => void
+  readonly doOpenEditor: () => void
+  readonly doCloseEditor: (projectId: backendModule.ProjectId) => void
 }
 
 /** The table of project assets. */
 export default function AssetsTable(props: AssetsTableProps) {
   const { hidden, query, setQuery, setProjectStartupInfo, setCanDownload, category } = props
-  const { setSuggestions, initialProjectName, projectStartupInfo } = props
+  const { setSuggestions, initialProjectName } = props
   const { assetListEvents, dispatchAssetListEvent, assetEvents, dispatchAssetEvent } = props
-  const { doOpenEditor: doOpenEditorRaw, doCloseEditor: doCloseEditorRaw } = props
+  const { doOpenEditor, doCloseEditor } = props
   const { setAssetPanelProps, targetDirectoryNodeRef, setIsAssetPanelTemporarilyVisible } = props
 
   const { user } = authProvider.useNonPartialUserSession()
@@ -421,7 +411,8 @@ export default function AssetsTable(props: AssetsTableProps) {
       backendModule.createRootDirectoryAsset(rootDirectoryId),
       rootParentDirectoryId,
       rootParentDirectoryId,
-      -1
+      -1,
+      backend.rootPath
     )
   })
   const [isDraggingFiles, setIsDraggingFiles] = React.useState(false)
@@ -893,7 +884,6 @@ export default function AssetsTable(props: AssetsTableProps) {
           dispatchAssetEvent({
             type: AssetEventType.openProject,
             id: projectToLoad.id,
-            shouldAutomaticallySwitchPage: true,
             runInBackground: false,
           })
         })
@@ -953,9 +943,16 @@ export default function AssetsTable(props: AssetsTableProps) {
           rootParentDirectoryId,
           rootParentDirectoryId,
           newAssets.map(asset =>
-            AssetTreeNode.fromAsset(asset, rootDirectory.id, rootDirectory.id, 0)
+            AssetTreeNode.fromAsset(
+              asset,
+              rootDirectory.id,
+              rootDirectory.id,
+              0,
+              `${backend.rootPath}/${asset.title}`
+            )
           ),
           -1,
+          backend.rootPath,
           rootDirectory.id,
           true
         )
@@ -974,7 +971,6 @@ export default function AssetsTable(props: AssetsTableProps) {
               dispatchAssetEvent({
                 type: AssetEventType.openProject,
                 id: projectToLoad.id,
-                shouldAutomaticallySwitchPage: true,
                 runInBackground: false,
               })
             })
@@ -995,7 +991,7 @@ export default function AssetsTable(props: AssetsTableProps) {
         return null
       })
     },
-    [rootDirectoryId, toastAndLog, setNameOfProjectToImmediatelyOpen, dispatchAssetEvent]
+    [rootDirectoryId, backend.rootPath, dispatchAssetEvent, toastAndLog]
   )
   const overwriteNodesRef = React.useRef(overwriteNodes)
   overwriteNodesRef.current = overwriteNodes
@@ -1006,40 +1002,29 @@ export default function AssetsTable(props: AssetsTableProps) {
     }
   }, [backend, category])
 
-  asyncEffectHooks.useAsyncEffect(
-    null,
-    async signal => {
-      setSelectedKeys(new Set())
-      try {
-        const newAssets = await backend
-          .listDirectory(
-            {
-              parentId: null,
-              filterBy: CATEGORY_TO_FILTER_BY[category],
-              recentProjects: category === Category.recent,
-              labels: null,
-            },
-            // The root directory has no name. This is also SAFE, as there is a different error
-            // message when the directory is the root directory (when `parentId == null`).
-            '(root)'
-          )
-          .catch(error => {
-            toastAndLog('listRootFolderBackendError', error)
-            throw error
-          })
-        if (!signal.aborted) {
-          setIsLoading(false)
-          overwriteNodes(newAssets)
-        }
-      } catch (error) {
-        if (!signal.aborted) {
-          setIsLoading(false)
-          toastAndLog(null, error)
-        }
-      }
-    },
-    [category, backend, setSelectedKeys]
+  const rootDirectoryQuery = backendHooks.useBackendQuery(
+    backend,
+    'listDirectory',
+    [
+      {
+        parentId: null,
+        filterBy: CATEGORY_TO_FILTER_BY[category],
+        recentProjects: category === Category.recent,
+        labels: null,
+      },
+      // The root directory has no name. This is also SAFE, as there is a different error
+      // message when the directory is the root directory (when `parentId == null`).
+      '(root)',
+    ],
+    { queryKey: [], staleTime: 0, meta: { persist: false } }
   )
+
+  React.useEffect(() => {
+    if (rootDirectoryQuery.data) {
+      setIsLoading(false)
+      overwriteNodes(rootDirectoryQuery.data)
+    }
+  }, [rootDirectoryQuery.data, overwriteNodes])
 
   React.useEffect(() => {
     const savedEnabledColumns = localStorage.get('enabledColumns')
@@ -1102,7 +1087,8 @@ export default function AssetsTable(props: AssetsTableProps) {
                         backendModule.createSpecialLoadingAsset(directoryId),
                         key,
                         directoryId,
-                        item.depth + 1
+                        item.depth + 1,
+                        ''
                       ),
                     ],
                   })
@@ -1144,7 +1130,13 @@ export default function AssetsTable(props: AssetsTableProps) {
                     }
                   }
                   const childAssetNodes = Array.from(childAssetsMap.values(), child =>
-                    AssetTreeNode.fromAsset(child, key, directoryId, item.depth + 1)
+                    AssetTreeNode.fromAsset(
+                      child,
+                      key,
+                      directoryId,
+                      item.depth + 1,
+                      `${item.path}/${child.title}`
+                    )
                   )
                   const specialEmptyAsset: backendModule.SpecialEmptyAsset | null =
                     (initialChildren != null && initialChildren.length !== 0) ||
@@ -1158,7 +1150,8 @@ export default function AssetsTable(props: AssetsTableProps) {
                             specialEmptyAsset,
                             key,
                             directoryId,
-                            item.depth + 1
+                            item.depth + 1,
+                            ''
                           ),
                         ]
                       : initialChildren == null || initialChildren.length === 0
@@ -1229,7 +1222,6 @@ export default function AssetsTable(props: AssetsTableProps) {
                   type: AssetEventType.openProject,
                   id: item.item.id,
                   runInBackground: false,
-                  shouldAutomaticallySwitchPage: true,
                 })
                 break
               }
@@ -1805,26 +1797,6 @@ export default function AssetsTable(props: AssetsTableProps) {
       queuedAssetListEventsRef.current.push(event)
     }
   })
-
-  const doOpenEditor = React.useCallback(
-    (
-      project: backendModule.ProjectAsset,
-      setProject: React.Dispatch<React.SetStateAction<backendModule.ProjectAsset>>,
-      switchPage: boolean
-    ) => {
-      doOpenEditorRaw(backend, project, setProject, switchPage)
-    },
-    [backend, doOpenEditorRaw]
-  )
-
-  const doCloseEditor = React.useCallback(
-    (project: backendModule.ProjectAsset) => {
-      if (project.id === projectStartupInfo?.projectAsset.id) {
-        doCloseEditorRaw(project)
-      }
-    },
-    [projectStartupInfo, doCloseEditorRaw]
-  )
 
   const doCopy = React.useCallback(() => {
     unsetModal()