+++ app/ide-desktop/lib/types/globals.d.ts
@@ -83,6 +83,15 @@ interface MenuApi {
     readonly setShowAboutModalHandler: (callback: () => void) => void
 }
 
+// ==================
+// === System API ===
+// ==================
+
+/** `window.systemApi` exposes functionality related to the operating system. */
+interface SystemApi {
+    readonly showItemInFolder: (fullPath: string) => void
+}
+
 // ====================
 // === Version Info ===
 // ====================
@@ -109,6 +118,7 @@ declare global {
         readonly authenticationApi: AuthenticationApi
         readonly navigationApi: NavigationApi
         readonly menuApi: MenuApi
+        readonly systemApi?: SystemApi
         readonly versionInfo?: VersionInfo
         toggleDevtools: () => void
     }
@@ -207,6 +217,8 @@ declare global {
 
             // @ts-expect-error The index signature is intentional to disallow unknown env vars.
             readonly IS_IN_PLAYWRIGHT_TEST?: `${boolean}`
+            // @ts-expect-error The index signature is intentional to disallow unknown env vars.
+            readonly PWDEBUG?: '1'
 
             // === Electron watch script variables ===