+++ CHANGELOG.md
@@ -20,6 +20,8 @@
   widget.][10337]
 - [Fixed issue where picking "<Numeric literal>" variant in some ports
   disallowed changing it again.][10337]
+- [Added click through on table and vector visualisation][10340] clicking on
+  index column will select row or value in seperate node
 
 [10064]: https://github.com/enso-org/enso/pull/10064
 [10179]: https://github.com/enso-org/enso/pull/10179
@@ -31,6 +33,7 @@
 [10310]: https://github.com/enso-org/enso/pull/10310
 [10327]: https://github.com/enso-org/enso/pull/10327
 [10337]: https://github.com/enso-org/enso/pull/10337
+[10340]: https://github.com/enso-org/enso/pull/10340
 
 #### Enso Standard Library
 
@@ -42,6 +45,8 @@
 - [Implemented `.cast` to and from `Decimal` columns for the in-memory
   database.][10206]
 - [Implemented fallback to Windows-1252 encoding for `Encoding.Default`.][10190]
+- [Added Table.duplicates component][10323]
+- [Renamed `Table.order_by` to `Table.sort`][10372]
 
 [debug-shortcuts]:
 
@@ -50,6 +55,8 @@
 [10130]: https://github.com/enso-org/enso/pull/10130
 [10206]: https://github.com/enso-org/enso/pull/10206
 [10190]: https://github.com/enso-org/enso/pull/10190
+[10323]: https://github.com/enso-org/enso/pull/10323
+[10372]: https://github.com/enso-org/enso/pull/10372
 
 <br/>![Release Notes](/docs/assets/tags/release_notes.svg)